+++ apps/sv/frontend/src/contexts/SvContext.tsx
@@ -4,14 +4,13 @@ import { Contract } from 'common-frontend/lib/utils/interfaces';
 
 import { CoinRules } from '@daml.js/canton-coin-0.1.0/lib/CC/Coin';
 import { SvcRules } from '@daml.js/svc-governance/lib/CN/SvcRules';
-import { ContractId } from '@daml/types';
 
 type SvUiState =
   | {
       svUser: string;
       svPartyId: string;
       svcPartyId: string;
-      coinRulesContractId: ContractId<CoinRules>;
+      coinRules: Contract<CoinRules>;
       svcRules: Contract<SvcRules>;
     }
   | undefined;
@@ -19,14 +18,14 @@ type SvUiState =
 export const useSvcInfos = (): UseQueryResult<SvUiState> => {
   const { getSvcInfo } = useSvClient();
   return useQuery({
-    queryKey: ['getSvcInfo', SvcRules],
+    queryKey: ['getSvcInfo', SvcRules, CoinRules],
     queryFn: async () => {
       const resp = await getSvcInfo();
       return {
         svUser: resp.svUser,
         svPartyId: resp.svPartyId,
         svcPartyId: resp.svcPartyId,
-        coinRulesContractId: resp.coinRulesContractId as ContractId<CoinRules>,
+        coinRules: Contract.decodeOpenAPI(resp.coinRules, CoinRules),
         svcRules: Contract.decodeOpenAPI(resp.svcRules, SvcRules),
       };
     },